scripts/deploy.py
#!/usr/bin/env python3
"""
Dify Application Deployment Script
Automates the deployment of Dify applications using configuration files.
"""

import os
import sys
import yaml
import argparse
import requests
from pathlib import Path
from typing import Dict, Any, Optional

class DifyDeployer:
    def __init__(self, config_path: str, environment: str = "development"):
        self.config_path = config_path
        self.environment = environment
        self.config = self.load_config()
        self.dify_config = self.load_dify_config()
        
    def load_config(self) -> Dict[str, Any]:
        """Load the main configuration file."""
        try:
            with open(self.config_path, 'r') as f:
                return yaml.safe_load(f)
        except FileNotFoundError:
            print(f"Configuration file not found: {self.config_path}")
            sys.exit(1)
        except yaml.YAMLError as e:
            print(f"Error parsing configuration file: {e}")
            sys.exit(1)
    
    def load_dify_config(self) -> Dict[str, Any]:
        """Load Dify-specific configuration for the environment."""
        env_config_path = f"config/environments/{self.environment}.yaml"
        try:
            with open(env_config_path, 'r') as f:
                return yaml.safe_load(f)
        except FileNotFoundError:
            print(f"Environment configuration not found: {env_config_path}")
            sys.exit(1)
        except yaml.YAMLError as e:
            print(f"Error parsing environment configuration: {e}")
            sys.exit(1)
    
    def deploy_application(self, app_path: str) -> bool:
        """Deploy a single application."""
        try:
            with open(app_path, 'r') as f:
                app_config = yaml.safe_load(f)
            
            app_name = app_config['application']['name']
            print(f"Deploying application: {app_name}")
            
            # Here you would implement the actual deployment logic
            # This is a placeholder for the Dify API integration
            success = self._deploy_to_dify(app_config)
            
            if success:
                print(f"✅ Successfully deployed {app_name}")
                return True
            else:
                print(f"❌ Failed to deploy {app_name}")
                return False
                
        except Exception as e:
            print(f"Error deploying application {app_path}: {e}")
            return False
    
    def _deploy_to_dify(self, app_config: Dict[str, Any]) -> bool:
        """Deploy application to Dify platform."""
        # This is a placeholder implementation
        # In a real scenario, you would:
        # 1. Authenticate with Dify API
        # 2. Create/update application
        # 3. Configure workflows
        # 4. Set up knowledge bases
        # 5. Configure API endpoints
        
        print("  - Creating application in Dify...")
        print("  - Configuring workflows...")
        print("  - Setting up knowledge base...")
        print("  - Configuring API endpoints...")
        
        # Simulate deployment success
        return True
    
    def deploy_all(self) -> bool:
        """Deploy all applications in the apps directory."""
        apps_dir = Path("apps")
        if not apps_dir.exists():
            print("Apps directory not found")
            return False
        
        app_files = list(apps_dir.glob("*.yaml"))
        if not app_files:
            print("No application configuration files found")
            return False
        
        print(f"Found {len(app_files)} applications to deploy")
        
        success_count = 0
        for app_file in app_files:
            if self.deploy_application(str(app_file)):
                success_count += 1
        
        print(f"\nDeployment Summary:")
        print(f"✅ Successful: {success_count}")
        print(f"❌ Failed: {len(app_files) - success_count}")
        
        return success_count == len(app_files)

def main():
    parser = argparse.ArgumentParser(description="Deploy Dify applications")
    parser.add_argument("--config", "-c", default="config/dify-config.yaml",
                       help="Path to main configuration file")
    parser.add_argument("--environment", "-e", default="development",
                       choices=["development", "staging", "production"],
                       help="Deployment environment")
    parser.add_argument("--app", "-a", help="Deploy specific application file")
    
    args = parser.parse_args()
    
    deployer = DifyDeployer(args.config, args.environment)
    
    if args.app:
        # Deploy specific application
        success = deployer.deploy_application(args.app)
        sys.exit(0 if success else 1)
    else:
        # Deploy all applications
        success = deployer.deploy_all()
        sys.exit(0 if success else 1)

if __name__ == "__main__":
    main()
